VSCode: Improve Python/C++ intellisense (per-folder settings)
The VSCode Python and C/C++ extensison do not work exceptionally
well when the configuration comes from the `.code-workspace` file:
sometimes a mix of settings is in effect when switching between
files from different folders (causing warnings or even python
interpreter crashes).

The relevant settings are now written directly and only to the
settings file of each folder.